+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [2.4.2] - Unpublished
 
+### Changed
+
+- Changed separatus model (issue #459).
+
 ### Fixed
 
 - Fix problem in optimisation for power/balance.